Restores QuickActionSheetReducer
