+++ app/src/main/java/org/mozilla/fenix/components/toolbar/BrowserStore.kt
@@ -3,9 +3,6 @@ package org.mozilla.fenix.components.toolbar
 import mozilla.components.lib.state.Action
 import mozilla.components.lib.state.State
 import mozilla.components.lib.state.Store
-import org.mozilla.fenix.quickactionsheet.QuickActionSheetAction
-import org.mozilla.fenix.quickactionsheet.QuickActionSheetState
-import org.mozilla.fenix.quickactionsheet.QuickActionSheetStore
 
 class BrowserStore(initialState: BrowserState) :
     Store<BrowserState, BrowserAction>(initialState, ::browserStateReducer)
@@ -25,37 +22,56 @@ data class BrowserState(
     val isAppLink: Boolean
 ) : State
 
+sealed class BrowserAction : Action
+
+// TODO: Ughhh this has to be in the same file as the BrowserAction definition is that fine?
+// TODO: This is because sealed classes have `private init`
 /**
  * Actions to dispatch through the [QuickActionSheetStore] to modify [QuickActionSheetState] through the reducer.
  */
-
-// TODO: Again these names feel weird when pulled into an abstraction...
-sealed class BrowserAction : Action {
-    data class BookmarkedStateChange(val bookmarked: Boolean) : BrowserAction()
-    data class ReadableStateChange(val readable: Boolean) : BrowserAction()
-    data class ReaderActiveStateChange(val active: Boolean) : BrowserAction()
-    data class AppLinkStateChange(val isAppLink: Boolean) : BrowserAction()
-    object BounceNeededChange : BrowserAction()
+sealed class QuickActionSheetAction : BrowserAction() {
+    data class BookmarkedStateChange(val bookmarked: Boolean) : QuickActionSheetAction()
+    data class ReadableStateChange(val readable: Boolean) : QuickActionSheetAction()
+    data class ReaderActiveStateChange(val active: Boolean) : QuickActionSheetAction()
+    data class AppLinkStateChange(val isAppLink: Boolean) : QuickActionSheetAction()
+    object BounceNeededChange : QuickActionSheetAction()
 }
 
 /**
- * Reduces [QuickActionSheetAction]s to update [QuickActionSheetState].
+ * Reducers for [BrowserStore].
+ *
+ * A reducer is a function that receives the current [BrowserState] and an [Action] and then returns a new
+ * [BrowserState].
  */
 fun browserStateReducer(
     state: BrowserState,
     action: BrowserAction
 ): BrowserState {
     return when (action) {
-        is BrowserAction.BookmarkedStateChange ->
-            state.copy(bookmarked = action.bookmarked)
-        is BrowserAction.ReadableStateChange ->
-            state.copy(readable = action.readable)
-        is BrowserAction.ReaderActiveStateChange ->
-            state.copy(readerActive = action.active)
-        is BrowserAction.BounceNeededChange ->
-            state.copy(bounceNeeded = true)
-        is BrowserAction.AppLinkStateChange -> {
-            state.copy(isAppLink = action.isAppLink)
+        is QuickActionSheetAction -> {
+            QuickActionSheetStateReducer.reduce(state, action)
+        }
+    }
+}
+
+
+/**
+ * Reduces [QuickActionSheetAction]s to update [BrowserState].
+ */
+internal object QuickActionSheetStateReducer {
+    fun reduce(state: BrowserState, action: QuickActionSheetAction): BrowserState {
+        return when (action) {
+            is QuickActionSheetAction.BookmarkedStateChange ->
+                state.copy(bookmarked = action.bookmarked)
+            is QuickActionSheetAction.ReadableStateChange ->
+                state.copy(readable = action.readable)
+            is QuickActionSheetAction.ReaderActiveStateChange ->
+                state.copy(readerActive = action.active)
+            is QuickActionSheetAction.BounceNeededChange ->
+                state.copy(bounceNeeded = true)
+            is QuickActionSheetAction.AppLinkStateChange -> {
+                state.copy(isAppLink = action.isAppLink)
+            }
         }
     }
 }
